% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clean_accent.R
\name{clean_accent}
\alias{clean_accent}
\title{Remove accents from text data}
\usage{
clean_accent(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
a character vector
}
\description{
This function removes accents from a, e, i, o, u, y, c, and n.  It also converts 'ae' and 'oe' single unicode characters to 2 ascii characters.
Note that it only currently works on lowercase text data.  See http://unicodelookup.com/
}

