% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/find_ngrams.R
\name{find_ngrams}
\alias{find_ngrams}
\title{Convert a list of tokens to ngrams}
\usage{
find_ngrams(dat, n, verbose = FALSE)
}
\arguments{
\item{dat}{a list of character vectors}

\item{n}{the number of n-grams to compute}

\item{verbose}{whether to print a progress bar}
}
\value{
a list of character vectors
}
\description{
This code takes a list of text vectors, and returns a list of text vectors, including n-grams
}
\examples{
find_ngrams(
  list(
    c('one'), c('sent', 'one'),
    c('this', 'is', 'sentence', 'two'),
    c('this', 'is', 'sentence', 'three', 'sentence', 'three'),
    c('finally', 'we', 'have', 'a', 'fourth', 'longer', 'sentence'),
    character(0),
    NULL,
    NA,
    NaN
  ),
  n=3,
  verbose=TRUE
)
}
\references{
\url{http://stackoverflow.com/questions/16489748/converting-a-list-of-tokens-to-n-grams}
\url{https://github.com/markvanderloo/stringdist/issues/39}
\url{https://gist.github.com/markvanderloo/9ae6a15f7d74a0159aec}
}

