% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/find_skip_grams.R
\name{find_skip_grams}
\alias{find_skip_grams}
\title{Convert a list of tokens to skip-grams}
\usage{
find_skip_grams(dat, n, verbose = FALSE)
}
\arguments{
\item{dat}{a list of character vectors}

\item{n}{the number of skip frams}

\item{verbose}{whether to print a progress bar}
}
\value{
a list of character vectors
}
\description{
This code takes a list of text vectors, and returns a list of skip-grams
}
\examples{
find_skip_grams(
  list(
    c('one'), c('sent', 'one'),
    c('this', 'is', 'sentence', 'two'),
    c('finally', 'we', 'have', 'a', 'third', 'longer', 'sentence'),
    character(0),
    NULL,
    NA,
    NaN
  ),
  n=3,
  verbose=TRUE
)
}
\references{
\url{http://stackoverflow.com/questions/16489748/converting-a-list-of-tokens-to-n-grams}
}

