% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/find_tokens.R
\name{find_tokens}
\alias{find_tokens}
\title{Convert a list of tokens to ngrams}
\usage{
find_tokens(x, split = " ", regex = FALSE)
}
\arguments{
\item{x}{A character vector}

\item{split}{The token to split on}

\item{regex}{whether to treat the split token as a regular expression.}
}
\value{
a list of character vectors
}
\description{
This code takes a list of text vectors, and returns a list of text vectors, including n-grams
}

