% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/text_to_sparse_matrix.R
\name{plot.textVectors}
\alias{plot.textVectors}
\title{Plot a text vectors object}
\usage{
\method{plot}{textVectors}(x, ...)
}
\arguments{
\item{x}{object of class textVectors}

\item{...}{ignored}
}
\description{
Plot a text vectors object
}

