% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/text_to_sparse_dictionary_matrix.R
\name{predict.sparse_dict_text_matrix}
\alias{predict.sparse_dict_text_matrix}
\title{Project a new character vector of data into the featurespace defined by a previous model's dictionary.  THIS DOCUMENTATION COULD USE SOME WORK!}
\usage{
\method{predict}{sparse_dict_text_matrix}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{an object of class sparse_dict_text_matrix}

\item{newdata}{a character vector of new data to make into a sparse matrix by the model$dictionary}

\item{...}{passed to text_to_sparse_dictionary_matrix}
}
\value{
a sparse_dict_text_matrix object
}
\description{
This code takes a vector of text, and applies a pre-defined text to sparse matrix mapping to it.
}
\examples{
x <- c(
  'i like this package written by zach mayer',
  'this package is so much fun',
  'thanks zach for writing it',
  'this package is the best package',
  'i want to give zach mayer a million dollars')
y <- 'this is a new sentence about how much i like this package written
by zach mayer it is a cool package and he is a cool dude'
a <- text_to_sparse_dictionary_matrix(x, dictionary=c('zac', 'may', 'writ'))
predict(a, y)
}
\references{
\url{http://stackoverflow.com/questions/22489996/r-grepl-quickly-match-multiple-strings-against-multiple-substrings-returning-a}
}

