% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sparseprcomp.R
\name{predict.sparseprcomp}
\alias{predict.sparseprcomp}
\title{Fast PCA using the irlba library for SVD}
\usage{
\method{predict}{sparseprcomp}(object, newdata, ...)
}
\arguments{
\item{object}{a sparse.prcomp object}

\item{newdata}{a sparse matrix of new data}

\item{...}{ignored}
}
\value{
a sparse matrix
}
\description{
This functions performs principle components analysis via SVD,
using the irlba library.  Note that it is not possible to
center and scale a sparse matrix.   Folding in the
eigenvalues might help compensate for this.
}

