% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/text_to_sparse_matrix.R
\name{print.textVectors}
\alias{print.textVectors}
\title{Print a text vectors object}
\usage{
\method{print}{textVectors}(x, ...)
}
\arguments{
\item{x}{object of class textVectors}

\item{...}{ignored}
}
\description{
Print a text vectors object
}

