% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sparseprcomp.R
\name{sparseprcomp}
\alias{sparseprcomp}
\title{Fast PCA using the irlba library for SVD}
\usage{
sparseprcomp(x, n = 5, retx = FALSE, fold_in_eigens = FALSE, ...)
}
\arguments{
\item{x}{a sparse matrix}

\item{n}{number of principle components to calculate}

\item{retx}{whether to return the principle components of the input data}

\item{fold_in_eigens}{whether to add the eigenvectors to the principle component rotations}

\item{...}{arguments passed to irlbda}
}
\value{
a sparse.prcomp object, which inherits from prcomp
}
\description{
This functions performs principle components analysis via SVD,
using the irlba library.  Note that it is not possible to
center and scale a sparse matrix.   Folding in the
eigenvalues might help compensate for this.
}

