% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/text_to_sparse_dictionary_matrix.R
\name{text_to_sparse_dictionary_matrix}
\alias{text_to_sparse_dictionary_matrix}
\title{Tokenize a vector of text and convert to a sparse matrix based on the presense or absense of a dictionary of substrings.}
\usage{
text_to_sparse_dictionary_matrix(x, dictionary, verbose = FALSE,
  fixed = TRUE, freqCutoff = 0, absCutoff = 0)
}
\arguments{
\item{x}{a character vector}

\item{dictionary}{the dictionary to lookup in the character vector}

\item{verbose}{whether to print a log while performing the operations.}

\item{fixed}{whether to treat the dictionary as-is (TRUE) or as regular expressions (FALSE).  TRUE is usually faster}

\item{freqCutoff}{columns below this pct frequency will be removed from the final object}

\item{absCutoff}{columns below this absolute frequency will be removed from the final object}
}
\value{
a sparse_dict_text_matrix object
}
\description{
This code takes a vector of text, uses grepl to detect each of a set of substrings within the text, and builts a sparse matrix
}
\examples{
x <- c(
  'i like this package written by zach mayer',
  'this package is so much fun',
  'thanks zach for writing it',
  'this package is the best package',
  'i want to give zach mayer a million dollars')
text_to_sparse_dictionary_matrix(x, dictionary=c('zac', 'may', 'writ'))
}
\references{
\url{http://stackoverflow.com/questions/22489996/r-grepl-quickly-match-multiple-strings-against-multiple-substrings-returning-a}
}

