% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/text_to_sparse_matrix.R
\name{text_to_sparse_matrix}
\alias{text_to_sparse_matrix}
\title{Tokenize a vector of text and convert to a sparse matrix. TODO: tf-idf is BROKEN!!}
\usage{
text_to_sparse_matrix(x, split_token = " ", verbose = FALSE,
  freqCutoff = 0, absCutoff = 0, tfidf = FALSE, bagofwords = NULL,
  spellcheck = FALSE, remove_stopwords = FALSE, stem = FALSE,
  ngrams = 1, skips = 0, stops = NULL)
}
\arguments{
\item{x}{a character vector}

\item{split_token}{token to use to split the text data.  If NULL, text will not be tokenized and will be treated as dummy variables.}

\item{verbose}{whether to print a log while performing the operations}

\item{freqCutoff}{columns below this pct frequency will be removed from the final object}

\item{absCutoff}{columns below this absolute frequency will be removed from the final object}

\item{tfidf}{whether to apply tfidf weighting.  NOTE THAT THIS WILL CREATE A DENSE MATRIX, WHICH IN MANY CASES IS BAD.}

\item{bagofwords}{input bagofwords to use to construct the final matrix}

\item{spellcheck}{if TRUE tokens will be spellchecked before they are stemmed}

\item{remove_stopwords}{if TRUE, english stopwords will be removed from the tokens}

\item{stem}{if TRUE the tokens will be stemmed, after tokenizing and before creating a matrix}

\item{ngrams}{If great than 1, n-grams of this degree will be added to the word bag}

\item{skips}{If great than 0, skips of this degree will be added to the word bag}

\item{stops}{Optional list of stopwords, otherwise a default list will be used.}
}
\value{
a sparse_text_matrix object
}
\description{
This code takes a vector of text, cleans it up, tokenizes it, spellchecks it,
removes stopwords, stems it, finds n-grams, crates a bag og words and
converts it to a sparse matrix using a bag of words model.  Optinally
it also applies td-idf weighting to the matrix.  This function can be slow.
Note that freqCutoff and absCutoff are relative to the number of documents
the term appears in, and ignore its frequency within documents.
}
\examples{
x <- c(
  'i like this package written by zach mayer',
  'this package is so much fun',
  'thanks zach for writing it',
  'this package is the best package',
  'i want to give zach mayer a million dollars')
text_to_sparse_matrix(
  x,
  absCutoff=1, ngrams=2, stem=TRUE, verbose=TRUE)
text_to_sparse_matrix(
  x,
  absCutoff=1, ngrams=2, skips=1, stem=TRUE, verbose=TRUE, tfidf=TRUE)
text_to_sparse_matrix(
  x,
  split_token=NULL, stem=TRUE, verbose=TRUE)
}
\references{
\url{http://stackoverflow.com/questions/4942361/how-to-turn-a-list-of-lists-to-a-sparse-matrix-in-r-without-using-lapply}
\url{http://en.wikipedia.org/wiki/Tf-idf}
}

