% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/text_normalization.R
\name{text_normalization}
\alias{text_normalization}
\title{Preprocess text data for use in PCA}
\usage{
text_normalization(x, remove = c("'"), spaces = c("[[:punct:]]",
  "[[:space:]]+"), remove_accents = FALSE, lowercase = TRUE)
}
\arguments{
\item{x}{a character vector}

\item{remove}{is a character vector (or regex) of characters to remove entirely}

\item{spaces}{is a character vector (or regex) of characters to convert to spaces}

\item{remove_accents}{boolean, decide whether or not to remove accents}

\item{lowercase}{Decide whether or not to convert the text to lowercase}
}
\value{
a character vector
}
\description{
Normalize text data by removing accents, special characters, and normalizing whitespace.
}

